+++ configs/eslint/js/react/index.js
@@ -1,0 +1,12 @@
+module.exports = {
+  extends: ['airbnb', 'airbnb/hooks', '../base'],
+  settings: {
+    react: {
+      version: 'detect'
+    }
+  },
+  rules: {
+    'react/jsx-filename-extension': 0,
+    'react/prop-types': 0
+  }
+}