+++ configs/eslint/js/react/test.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function FooComponent({ time, className }) {
+  return <div className={className}>FooComponent Component: {time}</div>
+}
+
+export default FooComponent