+++ scss/base/stylelint/index.js
@@ -1,6 +1,0 @@
-module.exports = {
-  extends: ['../../../css/base/stylelint', 'stylelint-config-standard-scss'],
-  rules: {
-    // override rules here
-  }
-}