+++ CHANGELOG.md
@@ -8,3 +8,18 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 * Initial proposal.
+
+## 2.0.0 (2023-03-24)
+
+### Added
+
+* Code-config CLI.
+* `eslint` command to create a new ESLint configuration automatically.
+* `stylelint` command to create a new Stylelint configuration automatically.
+* Auto-install necessary dependencies when running commands.
+* Documentation.
+
+### Updates
+
+* Moves the existing configuration to `config/` folder.
+* Inverts the language and tool in the folder structure.